backend/requirements.txt
tensorflow
librosa
numpy
pandas
scikit-learn
matplotlib
fastapi
uvicorn
python-multipart
python-dotenv
requests
gunicorn
